add file with interfaces
