+++ src/components/UserInfo/index.tsx
@@ -1,31 +1,5 @@
 import styles from './UserInfo.module.scss';
-
-interface User {
-    gender: string;
-    name: {
-        title: string;
-        first: string;
-        last: string;
-    };
-    dob: {
-        date: string;
-    };
-    location: {
-        postcode: string;
-        state: string;
-        city: string;
-        street: string;
-    };
-    phone: string;
-    picture: {
-        large: string;
-        medium: string;
-        thumbnail: string;
-    };
-    registered: {
-        date: string;
-    };
-}
+import { User } from '../../constans/interfaces';
 
 export const UserInfo = ({ user }: { user: User }) => {
     return (
