+++ src/pages/PageUserInfo/index.tsx
@@ -1,34 +1,9 @@
 import { useLocation } from 'react-router-dom';
 import { FormattedMessage } from 'react-intl';
+
 import { UserInfo } from '../../components';
 import styles from './PageUserInfo.module.scss';
-
-interface User {
-    gender: string;
-    name: {
-        title: string;
-        first: string;
-        last: string;
-    };
-    dob: {
-        date: string;
-    };
-    location: {
-        postcode: string;
-        state: string;
-        city: string;
-        street: string;
-    };
-    phone: string;
-    picture: {
-        large: string;
-        medium: string;
-        thumbnail: string;
-    };
-    registered: {
-        date: string;
-    };
-}
+import { User } from '../../constans/interfaces';
 
 const PageUserInfo = () => {
     const { state } = useLocation();
