+++ src/redux/users/reducer.ts
@@ -1,45 +1,5 @@
 import * as Type from './types';
-
-interface User {
-    gender: string;
-    name: {
-        title: string;
-        first: string;
-        last: string;
-    };
-    date: string;
-    location: {
-        postcode: string;
-        state: string;
-        city: string;
-        street: string;
-    };
-    phone: string;
-    picture: {
-        large: string;
-        medium: string;
-        thumbnail: string;
-    };
-    registered: {
-        date: string;
-    };
-    login: {
-        uuid: string;
-    };
-}
-
-interface UsersState {
-    users: User[];
-    currentPage: number;
-    isAuth: boolean;
-    isLoading: boolean;
-    error: Error | null | string;
-}
-
-interface Action<T> {
-    type: string;
-    payload: T;
-}
+import { UsersState, Action, Reducer } from '../../constans/interfaces';
 
 const initialState: UsersState = {
     users: [],
@@ -48,7 +8,6 @@ const initialState: UsersState = {
     isLoading: false,
     error: null,
 };
-type Reducer = User & User[] & Error & boolean & number;
 
 export const usersReducer = <T extends Reducer>(
     state: UsersState,
